Change enemy database clearing method
I believe (but can't tell for sure) that this fixes a strange bug
that was causing levels to fail to load (thus crashing the game)
under certain conditions.
The conditions I noticed were that the level had to have enemies
and that the levels that crashed tended to be fairly long.